requirements.txt
# SD-Studio Requirements
# Python 3.10+ required

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Stable Diffusion Core
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
safetensors>=0.4.0

# ControlNet
controlnet-aux>=0.0.7

# Image Processing
Pillow>=10.0.0

# GPU Monitoring
nvidia-ml-py3>=7.352.0

# Optional: For better performance
# xformers  # Uncomment for faster inference (requires compatible CUDA)
